refactor: One implementation of flags
