+++ packages/cozy-flags/src/FlagSwitcher.jsx
