+++ packages/cozy-flags/src/connect.jsx
