+++ packages/cozy-flags/src/flag-node.spec.js
@@ -3,6 +3,6 @@
  */
 
 import flag from './flag'
-import testFlagAPI from '../tests'
+import testFlagAPI from './tests'
 
 testFlagAPI(flag)
