+++ packages/cozy-flags/src/flag.spec.js
@@ -1,4 +1,4 @@
 import flag from './flag'
-import testFlagAPI from '../tests'
+import testFlagAPI from './tests'
 
 testFlagAPI(flag)
